+++ changelog.md
@@ -1,3 +1,8 @@
+#### 1.0.0-beta.24
+
+- Added `@QueryParameters` to `aom/openapi` decorators
+- Marked `@Parameters` as **deprecated**
+- Improved `parameterId` generation: used only endpoint class and property name
 #### 1.0.0-beta.23
 
 - Fixed bug with endpoint metadata definition